Use a 23K256 SRAM chip over SPI to store pixel data
Since even the MSP430G2553 only has 512B of RAM, external RAM is required if we want to store
pixel data for more than a certain number of LEDs.  The Microchip 23K256 can be accessed over
SPI, and provides 32KB of RAM.  The processor clock has been increased to 16MHz to compensate
for the extra work of sending data to/from a 23K256; copcar demo has been tweaked to suit too
as it ran far too fast if no extra delay was added on each frame.